fix(scatter-hist): KDE marginals line-only (kde_fill=False default) + suptitle inherits SCITEX_STYLE
- stx_scatter_hist: add kde_fill: bool = False (default = outline only, no
  fill under KDE marginals for readability); kde_fill=True restores the
  filled look. Per-group KDE curves still render in palette colours on both
  the x (top) and y (right) marginals; datetime-x handling unchanged.
- RecordingFigure.suptitle now inherits the active SCITEX_STYLE: defaults to
  fonts.suptitle_pt (set to 7 in SCITEX.yaml) and fontweight "normal" (not
  bold) when the caller passes neither. Explicit fontsize/size and
  fontweight/weight kwargs still win. SCITEX.yaml suptitle_pt 9 -> 7.
- Extract figure super-text methods (suptitle/supxlabel/supylabel/text/
  set_supxyt/set_supxytc) into FigureTextMixin (_wrappers/_figure_text.py)
  to keep _figure.py under the module size limit; public API unchanged.
